// src/lib/archetypes/types.ts
/**
 * Archetype type definitions
 */

export type ArchetypeId =
  // Base 12 archetypes
  | 'hero'
  | 'shadow'
  | 'sage'
  | 'innocent'
  | 'explorer'
  | 'creator'
  | 'caregiver'
  | 'lover'
  | 'jester'
  | 'everyman'
  | 'magician'
  | 'outlaw'
  // Hybrid 6 archetypes
  | 'strategic_warrior'
  | 'benevolent_leader'
  | 'visionary_artist'
  | 'wandering_rebel'
  | 'warm_flirt'
  | 'relatable_neighbor';